pyproject.toml
[tool.poetry]
name = "meme-ai"
version = "0.1.0"
description = ""
authors = ["resister-boy <hassanpumped17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
openai = "^1.57.4"
langchain = "^0.3.12"
langchain-core = "^0.3.25"
langchain-community = "^0.3.12"
langchainhub = "^0.1.21"
langgraph = "^0.2.59"
langsmith = "^0.2.3"
langchain-openai = "^0.2.12"
langchain-text-splitters = "^0.3.3"
langchain-chroma = "^0.1.4"
langchain-cohere = "^0.3.3"
langchain-milvus = "^0.1.7"
deepl = "^1.20.0"
pandas = "^2.2.3"
rank-bm25 = "^0.2.2"
pymupdf = "^1.25.1"
pdfplumber = "^0.11.4"
pypdf = "^5.1.0"
matplotlib = "^3.10.0"
pymupdf4llm = "^0.0.17"
notebook = "^7.3.1"
python-dotenv = "^1.0.1"
pydantic = "^2.10.3"
lxml = "^5.3.0"
pillow = "^11.0.0"
lark = "^1.2.2"
arxiv = "^2.1.3"
tiktoken = "^0.8.0"
tenacity = "^9.0.0"
pymilvus = "^2.5.0"
protobuf = "^5.29.1"
sqlalchemy = "^2.0.36"
google-search-results = "^2.4.2"
llama-index-core = "^0.12.5"
llama-parse = "^0.5.17"
llama-index-readers-file = "^0.4.1"
flashrank = "^0.2.9"
docx2txt = "^0.8"
seaborn = "^0.13.2"
grandalf = "^0.8"
mypy = "^1.13.0"
pinecone = "^5.4.2"
scikit-learn = "^1.6.0"
wikipedia = "^1.4.0"
faiss-cpu = "^1.9.0.post1"
beautifulsoup4 = "^4.12.3"
chromadb = "^0.5.23"
fastapi = "^0.115.6"
uuid = "^1.30"
tavily-python = "^0.5.0"
langchain-experimental = "^0.3.3"
pyupbit = "^0.2.34"
pyjwt = "^2.10.1"
ta = "^0.11.0"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
plotly = "^5.24.1"
schedule = "^1.2.2"
python-docx = "^1.1.2"
undetected-chromedriver = "^3.5.5"
seleniumbase = "^4.33.11"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
